+++ packages/http/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sonata-api/http
 
+## 0.0.3
+
+### Patch Changes
+
+- Fix bugs
+
 ## 0.0.2
 
 ### Patch Changes