+++ packages/node-http/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sonata-api/node-http
 
+## 0.0.2
+
+### Patch Changes
+
+- Fix bugs
+- Updated dependencies
+  - @sonata-api/http@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes