+++ packages/system/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sonata-api/system
 
+## 0.0.22
+
+### Patch Changes
+
+- Fix bugs
+
 ## 0.0.21
 
 ### Patch Changes